Move 'ContactRoles' type to services/aws/types
